# Fix and refactor row-reduce