Closes #2640 and refactors the row reductions to use `load_vector` instead of `cub`.

It also changes the loading size to be 128 bits and retunes a bit so there is a small speed-up (at least for row reduce simple). Will add some measurements below after CUDA finishes compiling the code. Finally, it removes the `BLOCK_DIM` template parameter which should help with compile times as well.