@angeloskath 2025-10-04
Microbenchmark on A100. I see some speedup and some slow-down I think I will do some more tuning and merge afterwards.

Before
```
sum_axis --size 32x512x512 --axis 0,2    0.12215209007263184
sum_axis --size 32x512x1024 --axis 0,2   0.2857553958892822
sum_axis --size 32x512x2048 --axis 0,2   0.7597212791442871
sum_axis --size 256x512x512 --axis 0,2   0.9653127193450928
sum_axis --size 256x512x1024 --axis 0,2  2.779815196990967
sum_axis --size 256x512x2048 --axis 0,2  5.837887763977051
sum_axis --size 4096x512 --axis 1        0.06255722045898438
sum_axis --size 4096x1024 --axis 1       0.08173346519470215
sum_axis --size 4096x2048 --axis 1       0.1889786720275879
sum_axis --size 4096x4096 --axis 1       0.4005749225616455
sum_axis --size 4096x8192 --axis 1       0.7776291370391846
sum_axis --size 4096x16384 --axis 1      1.5305595397949219
```

After
```
sum_axis --size 32x512x512 --axis 0,2    0.14006733894348145
sum_axis --size 32x512x1024 --axis 0,2   0.4020357131958008
sum_axis --size 32x512x2048 --axis 0,2   0.7963836193084717
sum_axis --size 256x512x512 --axis 0,2   1.515977144241333
sum_axis --size 256x512x1024 --axis 0,2  3.0236945152282715
sum_axis --size 256x512x2048 --axis 0,2  6.018394947052002
sum_axis --size 4096x512 --axis 1        0.06427288055419922
sum_axis --size 4096x1024 --axis 1       0.09265923500061035
sum_axis --size 4096x2048 --axis 1       0.20172882080078125
sum_axis --size 4096x4096 --axis 1       0.40026307106018066
sum_axis --size 4096x8192 --axis 1       0.7386627197265625
sum_axis --size 4096x16384 --axis 1      1.443312406539917

```